chore(auth): move requests namespace
- use singulars
